+++ src/app/lib/widgets/lfb-object/lfb-object.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { Subject } from 'rxjs';
 
 import { LfbObjectComponent } from './lfb-object.component';
-/*
 describe('LfbObjectComponent', () => {
   let component: LfbObjectComponent;
   let fixture: ComponentFixture<LfbObjectComponent>;
@@ -14,11 +14,22 @@ describe('LfbObjectComponent', () => {
 
     fixture = TestBed.createComponent(LfbObjectComponent);
     component = fixture.componentInstance;
+    (component as any).control = {
+      setErrors: () => {}
+    };
+    component.formProperty = {
+      schema: {
+        properties: {}
+      },
+      properties: {},
+      valueChanges: new Subject(),
+      errorsChanges: new Subject(),
+      getProperty: () => null
+    } as any;
     fixture.detectChanges();
   });
 
-  fit('should create', () => {
+  it('should create', () => {
     expect(component).toBeTruthy();
   });
 });
-*/